refactor : move domain, infra package QUZ-42
