+++ search-service/search-domain/src/main/kotlin/com/grepp/quizy/search/domain/quiz/Quiz.kt
@@ -1,4 +1,4 @@
-package com.grepp.quizy.domain.quiz
+package com.grepp.quizy.search.domain.quiz
 
 class Quiz(
     val id: Long,
